+++ src/seedu/addressbook/Main.java
@@ -7,6 +7,7 @@ import seedu.addressbook.commands.*;
 import seedu.addressbook.model.AddressBook;
 import seedu.addressbook.parser.Parser;
 import seedu.addressbook.storage.StorageFile;
+import seedu.addressbook.ui.TextUi;
 
 import java.io.*;
 import java.util.List;
